requirements.txt
Flask==0.11.1
Flask-Migrate==1.8.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1.0
psycopg2==2.6.2
Flask-Bcrypt==0.7.1
Faker==0.7.3
ipdb==0.10.2
